Refactor: reorganize project folder structure
