+++ src/components/DetailPage/CommentDisplay.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import supabase from '../supabase/supabase';
+import supabase from '../../supabase/supabase';
 import dayjs from 'dayjs';
 
 const CommentDisplay = () => {
